UUID -> ID id hiding model methods Editing/Batch editing

+++ src/main/java/rs/luka/android/studygroup/ui/singleitemactivities/AddCourseActivity.java
@@ -1,4 +1,4 @@
-package rs.luka.android.studygroup.activities.singleitemactivities;
+package rs.luka.android.studygroup.ui.singleitemactivities;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -8,15 +8,13 @@ import android.os.Bundle;
 import android.os.Environment;
 import android.provider.MediaStore;
 import android.support.design.widget.TextInputLayout;
-import android.support.v4.app.Fragment;
 import android.support.v4.app.NavUtils;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.CardView;
+import android.support.v7.widget.Toolbar;
 import android.view.KeyEvent;
-import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.inputmethod.EditorInfo;
 import android.widget.EditText;
 import android.widget.ImageView;
@@ -26,52 +24,70 @@ import java.io.File;
 
 import rs.luka.android.studygroup.R;
 import rs.luka.android.studygroup.Utils;
-import rs.luka.android.studygroup.io.Adder;
 import rs.luka.android.studygroup.io.Limits;
+import rs.luka.android.studygroup.model.Course;
+import rs.luka.android.studygroup.model.Group;
+import rs.luka.android.studygroup.ui.recyclers.GroupActivity;
 
 /**
  * Created by luka on 13.7.15..
  */
-public class AddCourseFragment extends Fragment {
+public class AddCourseActivity extends AppCompatActivity {
 
-    private static final int IDEAL_IMAGE_DIMENSION = 300;
-    private static final int INTENT_IMAGE = 0;
-    private static final File imageDir = new File(Environment.getExternalStorageDirectory().toString() + "/DCIM/StudyGroup/");
+    private static final int  IDEAL_IMAGE_DIMENSION = 300;
+    private static final int  INTENT_IMAGE          = 0;
+    private static final File imageDir              = new File(
+            Environment.getExternalStorageDirectory().toString() + "/DCIM/StudyGroup/");
 
-    private EditText subject;
+    private EditText        subject;
     private TextInputLayout subjectTil;
-    private EditText teacher;
+    private EditText        teacher;
     private TextInputLayout teacherTil;
-    private EditText year;
+    private EditText        year;
     private TextInputLayout yearTil;
-    private CardView add;
-    private ImageView image;
-    private File imageFile;
+    private CardView        add;
+    private ImageView       image;
+    private File            imageFile;
+    private Group           group;
+    private Course          course;
+    private boolean         editing;
 
-
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setHasOptionsMenu(true);
-    }
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_add_course, container, false);
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
 
-        AppCompatActivity ac = (AppCompatActivity) getActivity();
-        if (NavUtils.getParentActivityIntent(ac) != null) {
-            ac.getSupportActionBar().setDisplayHomeAsUpEnabled(true); //because reasons
+        group = getIntent().getParcelableExtra(GroupActivity.EXTRA_GROUP);
+        course = getIntent().getParcelableExtra(GroupActivity.EXTRA_COURSE);
+        editing = course != null;
+
+        setContentView(R.layout.activity_add_course);
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+
+        if (NavUtils.getParentActivityIntent(this) != null) {
+            getSupportActionBar().setDisplayHomeAsUpEnabled(true); //because reasons
         }
 
-        subject = (EditText) view.findViewById(R.id.add_course_name_input);
-        subjectTil = (TextInputLayout) view.findViewById(R.id.add_course_name_til);
-        teacher = (EditText) view.findViewById(R.id.add_course_prof_input);
-        teacherTil = (TextInputLayout) view.findViewById(R.id.add_course_prof_til);
-        year = (EditText) view.findViewById(R.id.add_course_year_input);
-        yearTil = (TextInputLayout) view.findViewById(R.id.add_course_year_til);
-        add = (CardView) view.findViewById(R.id.button_add);
-        image = (ImageView) view.findViewById(R.id.add_course_image);
+        subject = (EditText) findViewById(R.id.add_course_name_input);
+        subjectTil = (TextInputLayout) findViewById(R.id.add_course_name_til);
+        teacher = (EditText) findViewById(R.id.add_course_prof_input);
+        teacherTil = (TextInputLayout) findViewById(R.id.add_course_prof_til);
+        year = (EditText) findViewById(R.id.add_course_year_input);
+        yearTil = (TextInputLayout) findViewById(R.id.add_course_year_til);
+        add = (CardView) findViewById(R.id.button_add);
+        image = (ImageView) findViewById(R.id.add_course_image);
+        if (editing) {
+            subject.setText(course.getSubject());
+            teacher.setText(course.getTeacher());
+            if (course.getYear() != null) {
+                year.setText(course.getYear().toString());
+            }
+            if (course.hasImage()) {
+                image.setImageBitmap(course.getImage());
+            }
+            subject.setSelection(subject.getText().length());
+        }
 
         add.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -93,29 +109,28 @@ public class AddCourseFragment extends Fragment {
             @Override
             public void onClick(View v) {
                 Intent camera = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-                if (!imageDir.isDirectory())
-                    imageDir.mkdir();
-                imageFile = new File(imageDir, "course " + subject + " image.jpg");
-                if (imageFile.exists())
-                    imageFile.delete();
+                if (!imageDir.isDirectory()) { imageDir.mkdir(); }
+                imageFile = new File(imageDir, subject + " image.jpg");
+                if (imageFile.exists()) { imageFile.delete(); }
                 Intent gallery = new Intent(Intent.ACTION_PICK);
                 gallery.setType("image/*");
                 camera.putExtra(MediaStore.EXTRA_OUTPUT, Uri.fromFile(imageFile));
-                Intent chooserIntent = Intent.createChooser(camera, getString(R.string.select_image));
+                Intent chooserIntent = Intent.createChooser(camera,
+                                                            getString(R.string.select_image));
                 chooserIntent.putExtra(Intent.EXTRA_INITIAL_INTENTS, new Intent[]{gallery});
                 startActivityForResult(chooserIntent, INTENT_IMAGE);
             }
         });
 
-        return view;
     }
 
     @Override
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
         if (resultCode != Activity.RESULT_CANCELED) {
             if (requestCode == INTENT_IMAGE) {
-                if (data != null) { //ako je data==null, fotografija je napravljena kamerom, nije iz galerije
-                    imageFile = new File(Utils.getRealPathFromURI(getActivity(), data.getData()));
+                if (data
+                    != null) { //ako je data==null, fotografija je napravljena kamerom, nije iz galerije
+                    imageFile = new File(Utils.getRealPathFromURI(this, data.getData()));
                 }
                 BitmapFactory.Options opts = new BitmapFactory.Options();
                 opts.inJustDecodeBounds = true;
@@ -140,25 +155,29 @@ public class AddCourseFragment extends Fragment {
         } else if (subjectText.length() > Limits.COURSE_NAME_MAX_LENGTH) {
             subjectTil.setError(getString(R.string.error_too_long));
             error = true;
-        } else subjectTil.setError(null);
-        if (teacherText.length() > 128) {
+        } else { subjectTil.setError(null); }
+        if (teacherText.length() > Limits.COURSE_TEACHER_MAX_LENGTH) {
             teacherTil.setError(getString(R.string.error_too_long));
             error = true;
-        } else teacherTil.setError(null);
+        } else { teacherTil.setError(null); }
         if (!yearText.isEmpty() && Integer.parseInt(yearText) > Limits.COURSE_YEAR_MAX) {
             yearTil.setError(getString(R.string.error_number_too_large, getString(R.string.year)));
             error = true;
         } else if (!yearText.isEmpty() && Integer.parseInt(yearText) < Limits.COURSE_YEAR_MIN) {
             yearTil.setError(getString(R.string.error_negative_number, getString(R.string.year)));
             error = true;
-        } else yearTil.setError(null);
+        } else { yearTil.setError(null); }
         if (!error) {
-            if (imageFile.exists()) {
-                Adder.addCourse(subjectText, teacherText, yearText, imageFile);
+            if (editing) {
+                course.edit(subjectText, teacherText, yearText, imageFile);
             } else {
-                Adder.addCourse(subjectText, teacherText, yearText, null);
+                if (imageFile != null && imageFile.exists()) {
+                    group.addCourse(subjectText, teacherText, yearText, imageFile);
+                } else {
+                    group.addCourse(subjectText, teacherText, yearText, null);
+                }
             }
-            getActivity().onBackPressed();
+            onBackPressed();
         }
     }
 
@@ -166,7 +185,7 @@ public class AddCourseFragment extends Fragment {
     public boolean onOptionsItemSelected(MenuItem item) {
         switch (item.getItemId()) {
             case android.R.id.home:
-                NavUtils.navigateUpFromSameTask(getActivity());
+                NavUtils.navigateUpFromSameTask(this);
                 return true;
         }
         return super.onOptionsItemSelected(item);
